server/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
pydantic
python-dotenv
pandas
